start -> sender, end -> receiver
The current terminology is inconsistent.
The opposite of start is stop, not end.
The opposite of end is begin, not start.
sender and receiver is both symmetrical and
more accepted terminology for sequence diagrams.

The reason we do this change at all is that once
we support permalinks and store sequence diagrams in
databases, we are stuck with the terminology (unless
we want to go through a migration hell, which we don't).