Use mouseDebug search param to enable mouse overlay
According to ancient wisdom, compile-time flags are evil
because they cause problems. One such problem is that
our code coverage is lowered because of it. So make it
possible to enable it at runtime and add a test that
does so. This allows us to cach in sweet, sweet code
coverage %