v2.0.0
Abandon framgnet URLs in favor of integrated permalinks.
Since this breaks user API (old links won't work), bump
major version from 1.x.x to 2.x.x.

I don't think I have any users yet, and you can easily
restore a diagram (ask if you want to know how) using
if you have an old fragment URL.

I'm not taking it lightly to break links, and I expect
this to the be the only instance in the history of the
project. Not doing it now will create a lot of extra
work in the future. And again, I have not yet publicly
announced the project and I have not seen any traces
of users either so I think I can get away with it.