Validate JSON schema in aws-lambda-handler.js
AWS API Gateway does not support all JSON schema constructs
such as additionalProperties, so to get good JSON schema
validation we have to do it ourselves.

Since we don't want the aws-lambda-handler.js to be heavy
to execute we pre-pack the JSON schema validator with
webpack.

To make future code coverage work easier we don't bother
with minification. We can do it later if we see it makes
a big difference (which I don't think)