Upgrade to react-scripts@latest and stop using partial eject
Also adjust tests a bit to make them pass.
Also move them to src to enable react-scripts test.

In other words, this is a revert of 6eb3873b4f1e3c0 and then a bump
of react-scripts to the newest version. The old version we used is
more than 2 years old, i.e. ancient.